Moved Image Uploading to Components Folder
As we implement sending messages through socket.io we will need the same functionality so having this functionality in components only make sense
