+++ client/src/components/UploadImageForm/DropZone.tsx
